+++ changelog.md
@@ -5,6 +5,7 @@ vX.X.X
 - Added ability to close a calculator tab, closes #139.
 - Added support for function-based units for numerical calculator variables, closes #140.
 - Removed value property from the Unit class, closes #141.
+- Added "search" functionality to the calculator selection overlay, closes #142.
 
 v2.0.0
 ------